agentic_rag_app/requirements.txt
gradio>=4.0.0
llama-index>=0.10.0
llama-index-llms-litellm
llama-index-embeddings-litellm
llama-index-vector-stores-qdrant
qdrant-client>=1.7.0
litellm>=1.35.0
pyyaml>=6.0
python-dotenv>=1.0.0
asyncio
aiohttp
openai
anthropic
docling>=2.0.0
rerankers>=0.9.0
sentence-transformers>=2.0.0